feat(derive): compile subcommands from an enum
The largest gap in the derive: it described one command, so nothing with
subcommands could use it — which is every CLI it is meant for.

Three derives now cooperate. `Cli` on the root, `Subcommands` on an enum whose
variants each wrap a struct, and `Args` on those structs. A macro sees one
item, so they cannot share a counter or read each other's fields, and two
things follow from that.

Keys carry a hash of the type they came from, in the high half, with the index
in the low half and a tag for whether it is a command, a flag, or an argument.
Independently expanded macros therefore cannot hand two fields the same key,
which matters because a parse dispatches on it. Two type names would have to
collide in 32 bits, and `Spec::to_kdl` now asserts the tree holds no
duplicates, so that fails a test rather than binding the wrong field.

The tables are joined through two traits whose associated consts a parent
splices into its own `static` tables — consts rather than methods precisely
because a method call is not allowed in a `static`. Nothing is assembled at run
time, which was the point of the exercise.

Along the way, every derived struct now collects into a generated `Partial`
rather than into locals, since a subcommand's values cannot live in the root's
parse function. That replaced the prefixed-locals machinery entirely.

A doc comment on a variant overrides the struct's, because that is where a
reader of the enum expects to describe the command; ignoring it would lose the
description without saying so. Overriding one field of a const's metadata keeps
the tables static.

Eight tests, including one asserting keys are unique across separately expanded
types, and a snapshot of the emitted spec — nested `cmd` blocks that read like
something a person would have written.

+++ argv/src/lib.rs
@@ -105,7 +105,14 @@ pub struct Command<'a> {
     /// [`UnknownFlags`].
     pub unknown_flags: UnknownFlags,
     /// Caller-assigned identifier, echoed back in [`Event::Command`].
-    pub key: u32,
+    ///
+    /// Wide enough for a derive to make these unique without coordination: two
+    /// macro expansions cannot see each other, so the generated keys carry a hash
+    /// of the type they came from in the high half and a per-type index in the low
+    /// half. A parse dispatches on this, so a collision would bind the wrong field
+    /// — [`Spec::to_kdl`](crate::spec::Spec::to_kdl) checks the tree for duplicates
+    /// in debug builds.
+    pub key: u64,
 }
 
 impl Command<'_> {
@@ -126,7 +133,8 @@ impl Command<'_> {
 pub struct Flag<'a> {
     /// Caller-assigned identifier, echoed back in [`Event::Flag`]. This is how
     /// generated code knows which field to assign without any string comparison.
-    pub key: u32,
+    /// See [`Command::key`] on why it is this wide.
+    pub key: u64,
     /// Unused by binding, kept so a table entry can carry its own name for
     /// diagnostics.
     pub name: &'a str,
@@ -175,8 +183,9 @@ impl Flag<'_> {
 /// A positional argument.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub struct Arg<'a> {
-    /// Caller-assigned identifier, echoed back in [`Event::Arg`].
-    pub key: u32,
+    /// Caller-assigned identifier, echoed back in [`Event::Arg`]. See
+    /// [`Command::key`] on why it is this wide.
+    pub key: u64,
     /// Whether this argument keeps taking values once it has one.
     pub var: bool,
     /// This argument's relationship to the `--` separator.
